PDE condition implemented
