--- DeepINN/constraint/physics_loss.py
@@ -1,1 +1,0 @@
-pass
