+++ DeepINN/constraint/templates.py
@@ -1,0 +1,10 @@
+import torch
+
+def burgers_pde(X, y):
+    """
+    The burgers equation.
+    """
+    dy_x = dde.grad.jacobian(y, x, i=0, j=0)
+    dy_t = dde.grad.jacobian(y, x, i=0, j=1)
+    dy_xx = dde.grad.hessian(y, x, i=0, j=0)
+    return dy_t + y * dy_x - 0.01 / np.pi * dy_xx
